create day and smoothie type
