+++ src/app/smoothie/smoothie.action.js
@@ -6,28 +6,6 @@ import {
   CLOSE_SWITCHER,
 } from '../constants';
 
-export type Conference = {
-  'day': string,
-  'description': string,
-  'domain': string,
-  'duration': number,
-  'level': string,
-  'room': string,
-  'speaker': {
-    'city': string,
-    'company': string,
-    'gender': string,
-    'linkedinLink': string,
-    'name': string,
-    'pictureLink': string,
-    'twitterLink': string,
-  },
-  'sponsored': boolean,
-  'tags': Array<string>,
-  'timeBegin': string,
-  'title': string,
-};
-
 export const addConferenceAction = conferences => ({
   type: MIX_CONFERENCES,
   data: {
